service-ai-workflow/pyproject.toml
[project]
name = "backend-ai"
version = "0.1.0"
description = "AI Agent for PR reviews and automated fixes"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn>=0.27.1",
    "fastmcp>=0.4.1",
    "python-dotenv>=1.0.1",
    "requests>=2.31.0",
    "openai>=1.12.0",
    "anthropic>=0.18.0",
    "pydantic>=2.6.1",
    "pydantic-settings>=2.2.1",
    "python-multipart>=0.0.9",
    "gitpython>=3.1.41",
    "elasticsearch>=8.0.0",
    "groq>=0.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
